safari-600.7-branch/LayoutTests/css3/flexbox/flex-property-parsing-expected.txt

Tests setting the -webkit-flex propery.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS flexitem.style.webkitFlex is ""
PASS getComputedStyle(flexitem).webkitFlex is "0 1 auto"
PASS flexitem.style.webkitFlex is ""
PASS getComputedStyle(flexitem).webkitFlex is "0 1 auto"
PASS flexitem.style.webkitFlex is ""
PASS getComputedStyle(flexitem).webkitFlex is "0 1 auto"
PASS flexitem.style.webkitFlex is "2 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "2 1 0px"
PASS flexitem.style.webkitFlex is "0 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 1 0px"
PASS flexitem.style.webkitFlex is "1.5 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "1.5 1 0px"
PASS flexitem.style.webkitFlex is "1 1 auto"
PASS getComputedStyle(flexitem).webkitFlex is "1 1 auto"
PASS flexitem.style.webkitFlex is "1 1 1px"
PASS getComputedStyle(flexitem).webkitFlex is "1 1 1px"
PASS flexitem.style.webkitFlex is "1 1 2em"
PASS getComputedStyle(flexitem).webkitFlex is "1 1 32px"
PASS flexitem.style.webkitFlex is "1 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "1 1 0px"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 0px"
PASS flexitem.style.webkitFlex is "0 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 1 0px"
PASS flexitem.style.webkitFlex is "1 0 0px"
PASS getComputedStyle(flexitem).webkitFlex is "1 0 0px"
PASS flexitem.style.webkitFlex is "2 1 auto"
PASS getComputedStyle(flexitem).webkitFlex is "2 1 auto"
PASS flexitem.style.webkitFlex is "3 1 4px"
PASS getComputedStyle(flexitem).webkitFlex is "3 1 4px"
PASS flexitem.style.webkitFlex is "5.25 1 auto"
PASS getComputedStyle(flexitem).webkitFlex is "5.25 1 auto"
PASS flexitem.style.webkitFlex is "4 1 6em"
PASS getComputedStyle(flexitem).webkitFlex is "4 1 96px"
PASS flexitem.style.webkitFlex is "4 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "4 1 0px"
PASS flexitem.style.webkitFlex is "0 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 1 0px"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "1 2 0px"
PASS getComputedStyle(flexitem).webkitFlex is "1 2 0px"
PASS flexitem.style.webkitFlex is "0 0 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 0px"
PASS flexitem.style.webkitFlex is "1 2 auto"
PASS getComputedStyle(flexitem).webkitFlex is "1 2 auto"
PASS flexitem.style.webkitFlex is "1.75 2 3px"
PASS getComputedStyle(flexitem).webkitFlex is "1.75 2 3px"
PASS flexitem.style.webkitFlex is "1 2 3px"
PASS getComputedStyle(flexitem).webkitFlex is "1 2 3px"
PASS flexitem.style.webkitFlex is "1 2 3px"
PASS getComputedStyle(flexitem).webkitFlex is "1 2 3px"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 1 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 1 auto"
PASS flexitem.style.webkitFlex is "3 1 auto"
PASS getComputedStyle(flexitem).webkitFlex is "3 1 auto"
PASS flexitem.style.webkitFlex is "0 1 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 1 0px"
PASS flexitem.style.webkitFlex is "0 0 0px"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 0px"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS flexitem.style.webkitFlex is "initial"
PASS getComputedStyle(flexitem).webkitFlex is "0 1 auto"
PASS flexitem.style.webkitFlex is "0 0 auto"
PASS getComputedStyle(flexitem).webkitFlex is "0 0 auto"
PASS successfullyParsed is true

TEST COMPLETE
